Conflicts:
	thebes-code/hat/client/src/main/java/edu/berkeley/thebes/hat/client/ThebesHATClient.java
	thebes-code/hat/server/src/main/java/edu/berkeley/thebes/hat/server/dependencies/DependencyResolver.java
	thebes-code/hat/server/src/main/java/edu/berkeley/thebes/hat/server/dependencies/PendingWrites.java
	thebes-code/twopl/tm/src/main/java/edu/berkeley/thebes/twopl/tm/TwoPLTransactionClient.java

+++ thebes-code/twopl/tm/src/main/java/edu/berkeley/thebes/twopl/tm/TwoPLTransactionClient.java
@@ -67,7 +67,7 @@ public class TwoPLTransactionClient implements IThebesClient {
         }
         
         long timestamp = System.currentTimeMillis();
-        DataItem dataItem = new DataItem(value, new Version((short) clientId, timestamp));
+        DataItem dataItem = new DataItem(value, new Version(clientId, timestamp));
         return masterRouter.getMasterByKey(key).put(sessionId, key, DataItem.toThrift(dataItem));
     }
